Adding rating bar

+++ IntelligentHabitacion.App/Model/ScheduleUserModel.cs
@@ -15,7 +15,7 @@ namespace IntelligentHabitacion.App.Model
     {
         public DateTime Date { get; set; }
         public string Room { get; set; }
-        public short RatingStars { get; set; }
+        public sbyte RatingStars { get; set; }
         public bool CanBeRate { get; set; }
     }
 }
